Major refactoring for entry
